ContentCreationRevision.DjangoAPI/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
azure-ai-formrecognizer==3.3.3
azure-ai-ml==1.24.0
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.4
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-search-documents==11.5.2
azure-storage-blob==12.24.1
azure-storage-file-datalake==12.18.1
azure-storage-file-share==12.20.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.0
dataclasses-json==0.6.7
deepsearch-glm==1.0.0
Deprecated==1.2.16
dill==0.3.9
distro==1.9.0
django-cors-headers==4.6.0
Django==5.1.5
djangorestframework==3.15.2
docling==2.13.0
docling-core==2.15.1
docling-ibm-models==3.3.0
docling-parse==3.1.1
drf-spectacular==0.28.0
durationpy==0.9
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.115.7
filelock==3.17.0
filetype==1.2.0
fixedint==0.1.6
flatbuffers==25.1.21
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langsmith==0.3.1
lazy_loader==0.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-dbapi==0.50b0
opentelemetry-instrumentation-django==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-psycopg2==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-urllib==0.50b0
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-instrumentation-wsgi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pip==24.0
portalocker==2.10.1
posthog==3.9.3
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.7.1
pydash==8.0.5
Pygments==2.19.1
PyJWT==2.10.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.1
semchunk==2.2.2
setuptools==70.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.0
starlette==0.45.3
strictyaml==1.7.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.1.10
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
typer==0.12.5
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
whitenoise==6.8.2
wrapt==1.17.2
XlsxWriter==3.2.1
yarl==1.18.3
zipp==3.20.2
zstandard==0.23.0
